+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 - Deprecation warning for visibility of `ClassiEtaSpecieNISECI::{specie, new}`
 - Deprecation warning for definition of `domain::hfbi::RIFERIMENTO_HFBI`
 - Deprecation warning for visibility of `domain::niseci::ValoriIntermediNISECI::{log, to_csv}`
+- Deprecation warning for visibility of `domain::niseci::EsemplariPerCattura` fields
 
 ## [0.1.5] - 2026-04-30